Custom Consistent Step Layouts (#16)
Add a layout component prop to wizard. Which moves the rendering of buttons from the lib to the user. This layout component will call props.children to render the step and props[props.scopeKey].buttons to render the buttons. Layout component receives props from the top level wizard and step specific instances.

scopeKey used for any props that are provided by the wizard to your components, any prop that is from a prop of props do not have scope.